refactor(v0): extract plan-session service
